+++ CHANGELOG.md
@@ -34,8 +34,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added bin:cli.js
 - Added multiple package outputs
 
-## [1.0.3] - 2020-01-?
+## [1.0.3] - 2020-01-11
 
 ### Fixed
 
 - Fixed mistakes in README
+- Fixed Jimp.read not found in standalone
+- Catching error while processing posts